fix test with typescript
