+++ package.json
@@ -14,9 +14,9 @@
   "license": "ISC",
   "dependencies": {
     "@anthropic-ai/sdk": "^0.78.0",
-    "@sentry/node": "^8.0.0",
     "@google-cloud/firestore": "^7.11.0",
     "@google/generative-ai": "^0.24.1",
+    "@sentry/node": "^8.0.0",
     "@types/cors": "^2.8.17",
     "@types/express": "^5.0.0",
     "@types/morgan": "^1.9.9",
@@ -38,6 +38,7 @@
     "@babel/preset-env": "^7.26.0",
     "@babel/preset-typescript": "^7.28.5",
     "@eslint/js": "^9.14.0",
+    "@types/jest": "^30.0.0",
     "@types/swagger-jsdoc": "^6.0.4",
     "@types/swagger-ui-express": "^4.1.8",
     "babel-jest": "^29.7.0",
